refactor External stake view into a separate component & remove from Ohm Stake View
