+++ src/views/Stake/Stake.jsx
@@ -10,15 +10,8 @@ import {
   InputLabel,
   OutlinedInput,
   Button,
-  SvgIcon,
   Tab,
   Tabs,
-  TableHead,
-  TableCell,
-  TableBody,
-  Table,
-  TableRow,
-  TableContainer,
   Link,
   Zoom,
 } from "@material-ui/core";
@@ -26,18 +19,15 @@ import NewReleases from "@material-ui/icons/NewReleases";
 import CheckCircleIcon from "@material-ui/icons/CheckCircle";
 import RebaseTimer from "../../components/RebaseTimer/RebaseTimer";
 import TabPanel from "../../components/TabPanel";
-import BondLogo from "../../components/BondLogo";
 import { trim, getTokenImage, getOhmTokenImage } from "../../helpers";
 import { changeStake, changeApproval } from "../../slices/StakeThunk";
-import { getLusdData } from "../../slices/LusdSlice";
 import useMediaQuery from "@material-ui/core/useMediaQuery";
-import { ReactComponent as ArrowUp } from "../../assets/icons/arrow-up.svg";
 import "./stake.scss";
 import { NavLink } from "react-router-dom";
 import { useWeb3Context } from "src/hooks/web3Context";
 import { isPendingTxn, txnButtonText } from "src/slices/PendingTxnsSlice";
 import { Skeleton } from "@material-ui/lab";
-import OhmLusdImg from "src/assets/tokens/OHM-LUSD.svg";
+import ExternalStakePool from "./ExternalStakePool";
 
 function a11yProps(index) {
   return {
@@ -62,9 +52,6 @@ function Stake() {
   const isAppLoading = useSelector(state => state.app.loading);
   const currentIndex = useSelector(state => {
     return state.app.currentIndex;
-  });
-  const lusdData = useSelector(state => {
-    return state.lusdData;
   });
   const fiveDayRate = useSelector(state => {
     return state.app.fiveDayRate;
@@ -128,14 +115,6 @@ function Stake() {
     },
     [stakeAllowance],
   );
-
-  const loadLusdData = async () => {
-    await dispatch(getLusdData({ address: address, provider: provider, networkID: chainID }));
-  };
-
-  useEffect(() => {
-    loadLusdData();
-  }, [provider]);
 
   let modalButton = [];
 
@@ -399,111 +378,8 @@ function Stake() {
         </Paper>
       </Zoom>
 
-      <Zoom in={true}>
-        <Paper className={`ohm-card secondary ${isSmallScreen && "mobile"}`}>
-          <div className="card-header">
-            <Typography variant="h5">Farm Pool</Typography>
-          </div>
-          <div className="card-content">
-            {!isSmallScreen ? (
-              <TableContainer className="stake-table">
-                <Table>
-                  <TableHead>
-                    <TableRow>
-                      <TableCell>Asset</TableCell>
-                      <TableCell align="left">APR</TableCell>
-                      <TableCell align="left">TVL</TableCell>
-                      <TableCell align="left">Balance</TableCell>
-                      <TableCell></TableCell>
-                    </TableRow>
-                  </TableHead>
-
-                  <TableBody>
-                    <TableRow>
-                      <TableCell>
-                        <Box className="ohm-pairs">
-                          <BondLogo bond={{ bondIconSvg: OhmLusdImg, isLP: true }}></BondLogo>
-                          <Typography>OHM-LUSD</Typography>
-                        </Box>
-                      </TableCell>
-                      <TableCell align="left">{lusdData && trim(lusdData.apy, 1)}%</TableCell>
-                      <TableCell align="left">
-                        {lusdData &&
-                          lusdData.tvl &&
-                          new Intl.NumberFormat("en-US", {
-                            style: "currency",
-                            currency: "USD",
-                            maximumFractionDigits: 0,
-                            minimumFractionDigits: 0,
-                          }).format(lusdData.tvl)}
-                      </TableCell>
-                      <TableCell align="left"> {(lusdData && trim(lusdData.balance, 2)) || 0} LP </TableCell>
-                      <TableCell align="center">
-                        {/* TODO (appleseed-lusd): replace button link */}
-                        <Button
-                          variant="outlined"
-                          color="secondary"
-                          href="https://app.pickle.finance/farms"
-                          target="_blank"
-                          className="stake-lp-button"
-                        >
-                          <Typography variant="body1">Stake on Pickle</Typography>
-                          <SvgIcon component={ArrowUp} color="primary" />
-                        </Button>
-                      </TableCell>
-                    </TableRow>
-                  </TableBody>
-                </Table>
-              </TableContainer>
-            ) : (
-              <div className="stake-pool">
-                <div className={`pool-card-top-row ${isMobileScreen && "small"}`}>
-                  <Box className="ohm-pairs">
-                    <BondLogo bond={ohm_lusd}></BondLogo>
-                    <Typography gutterBottom={false}>OHM-LUSD</Typography>
-                  </Box>
-                </div>
-                <div className="pool-data">
-                  <div className="data-row">
-                    <Typography>APR</Typography>
-                    <Typography>{lusdData && trim(lusdData.apy, 1)}%</Typography>
-                  </div>
-                  <div className="data-row">
-                    <Typography>TVL</Typography>
-                    <Typography>
-                      {lusdData &&
-                        lusdData.tvl &&
-                        new Intl.NumberFormat("en-US", {
-                          style: "currency",
-                          currency: "USD",
-                          maximumFractionDigits: 0,
-                          minimumFractionDigits: 0,
-                        }).format(lusdData.tvl)}
-                    </Typography>
-                  </div>
-                  <div className="data-row">
-                    <Typography>Balance</Typography>
-                    <Typography>{(lusdData && trim(lusdData.balance, 2)) || 0} LP</Typography>
-                  </div>
-
-                  {/* TODO (appleseed-lusd): replace button link */}
-                  <Button
-                    variant="outlined"
-                    color="secondary"
-                    href="https://app.pickle.finance/farms"
-                    target="_blank"
-                    className="stake-lp-button"
-                    fullWidth
-                  >
-                    <Typography variant="body1">Stake on Pickle</Typography>
-                    <SvgIcon component={ArrowUp} color="primary" />
-                  </Button>
-                </div>
-              </div>
-            )}
-          </div>
-        </Paper>
-      </Zoom>
+      {/* TODO (appleseed-lusd): hiding Stake Pool temporarily, when ready update button links */}
+      {/* <ExternalStakePool /> */}
     </div>
   );
 }
